\name{Dpit-package}
\alias{Dpit-package}
\docType{package}
\title{Distribution Pitting}

\description{
Compares distributions with one another in terms of their fit to each sample in a dataset that contains multiple samples, as described in Joo, Aguinis, and Bradley (2017). Users can examine the fit of seven distributions per sample: pure power law, lognormal, exponential, power law with an exponential cutoff, normal, Poisson, and Weibull.
			}

\details{
Based on the R code found at:
\url{http://tuvalu.santafe.edu/~aaronc/powerlaws/}.  
In particular, we borrowed heavily from Cosma R. Shalizi's code. Finally, we owe much gratitude to Cosma R. Shalizi, Aaron Clauset, and Yogesh Virkar for their kind responses to our questions regarding the code we borrowed from.
		}

\author{
Harry Joo,
Herman Aguinis,
Kyle J. Bradley

Maintainer: Harry Joo <harryjoo19@gmail.com>
		}

\references{
Joo, H., Aguinis, H., & Bradley, K. J. 2017. Not all nonnormal distributions are created equal: Improved theoretical and measurement precision. Journal of Applied Psychology. Advance online publication. doi: 10.1037/apl0000214
			}

\keyword{package}