\name{FourSamples}
\alias{FourSamples}
\docType{data}
\title{
A part of the data set used in Joo, Aguinis, & Bradley (2017).
}
\description{
A part of the data set used in Joo, Aguinis, & Bradley (2017).
}
\usage{data(FourSamples)}
\format{
  A data frame consisting of four variables with varying number of observations.
  \describe{
    \item{\code{v1}}{number of publications in top five journals in the field of Agriculture.}
    \item{\code{v2}}{number of publications in top five journals in the field of Agronomy.}
    \item{\code{v228}}{number electrical fixtures assembled by assemblers.}
    \item{\code{v229}}{wirer's ratio of production time per unit assembled to standard.}
  }
}

\details{
Running the Dpit() function on the FourSamples dataset will produce the results reported in Table 4 in Joo, Aguinis, & Bradley (2017). Though the precise results the user obtains may very slightly differ from those in Joo et al.'s Table 4 due to statistical fluctuations, substantive conclusions remain the same regardless of such fluctuations.  
}

\author{
Harry Joo,
Herman Aguinis,
Kyle J. Bradley

Maintainer: Harry Joo <harryjoo19@gmail.com>
}
\source{
Joo, H., Aguinis, H., & Bradley, K. J. 2017. Not all nonnormal distributions are created equal: Improved theoretical and measurement precision. Journal of Applied Psychology. Advance online publication. doi: 10.1037/apl0000214
}
\keyword{datasets}
